requirements.txt
aws-cdk-lib==2.135.0
constructs>=10.0.0,<11.0.0
docker==6.1.3